requirements.txt

torch
torchvision
numpy
pandas
matplotlib
scikit-learn
opencv-python
tqdm
PyYAML
transformers
